Move NoteDataSource.kt
